Move all db managers to diferents folders, and add joins to verify if user own cart
