+++ M2/Back-end/pet_shop/db/sell_module/db_receipt_manager.py
@@ -1,7 +1,7 @@
 from sqlalchemy import insert, select, delete, update
-from .tables_manager import TablesManager
+from db.tables_manager import TablesManager
 from utils.api_exception import APIException
-from utils.helpers import filter_locals
+from utils.helpers import filter_locals, verify_user_own_cart
 from datetime import datetime
 
 receipt_table = TablesManager.receipt_table
@@ -25,21 +25,36 @@ class DbReceiptManager:
     def get_receipt(
         self,
         id: int | None = None,
+        id_user: int | None = None,
         id_cart: int | None = None,
         id_address: int | None = None,
         id_product: int | None = None,
         entry_date: str | None = None,
         state: str | None = None,
     ):
-        params = filter_locals(locals())
+        filters = filter_locals(
+            locals(),
+            (
+                "self",
+                "id_user",
+            ),
+        )
         conditions = []
-        for key, value in params.items():
-            if value is not None:
-                conditions.append(getattr(receipt_table.c, key) == value)
-        stmt = select(receipt_table)
-        if conditions:
-            stmt = stmt.where(and_(*conditions))
         with engine.connect() as conn:
+            if id_user is not None and not verify_user_own_cart(
+                conn, id, id_user, receipt_table
+            ):
+                raise APIException(
+                    f"Receipt id:{id} not owned by user id:{id_user}", 403
+                )
+            if entry_date is not None:
+                __validate_date_str(entry_date, "Entry date in receipt is not valid")
+            for key, value in filters.items():
+                if value is not None:
+                    conditions.append(getattr(receipt_table.c, key) == value)
+            stmt = select(receipt_table)
+            if conditions:
+                stmt = stmt.where(and_(*conditions))
             result = conn.execute(stmt).mappings().all()
         return result
 
@@ -55,49 +70,40 @@ class DbReceiptManager:
     ):
         if entry_date:
             __validate_date_str(entry_date, "Entry date in receipt is not valid")
-        params = filter_locals(locals(), ("self", "id", "id_user"))
+        values = filter_locals(locals(), ("self", "id", "id_user"))
 
         stmt = update(receipt_table).where(receipt_table.c.id == id)
         if values:
             stmt = stmt.values(**values)
-        # hacer join en cart para obtener user    
+        # hacer join en cart para obtener user
         with engine.connect() as conn:
             result = conn.execute(stmt)
             rows_updated = result.rowcount
-        if rows_updated != 0:
-            conn.commit()
-        raise APIException(
-            (
-                f"Recipt id:{str(id)} not exist or not owned by user id:{id_user}"
-                if id_user
-                else f"Recipt id:{str(id)} not exist"
-            ),
-            404,
-        )
+        if rows_updated == 0:
+            raise APIException(
+                (
+                    f"Recipt id:{str(id)} not exist or not owned by user id:{id_user}"
+                    if id_user
+                    else f"Recipt id:{str(id)} not exist"
+                ),
+                404,
+            )
+        conn.commit()
 
-    def delete_cart(self, id: int, id_user: int | None = None):
-        conditions = [receipt_table.c.id == id]
-        # Hacer join con Cart para verificar el user
-        if id_user is not None:
-            conditions.append(receipt_table.c.id_user == id_user)
-        stmt = delete(receipt_table).where(*conditions)
+    def delete_receipt(self, id: int):
+        stmt = delete(receipt_table).where(receipt_table.c.id == id)
         with engine.connect() as conn:
             result = conn.execute(stmt)
             rows_deleted = result.rowcount
-        if rows_deleted != 0:
-            conn.commit()
-        raise APIException(
-            (
-                f"Recipt id:{str(id)} not exist or not owned by user id:{id_user}"
-                if id_user
-                else f"Recipt id:{str(id)} not exist"
-            ),
-            404,
-        )
+        if rows_deleted == 0:
+            raise APIException(
+                (f"Recipt id:{str(id)} not exist"),
+                404,
+            )
+        conn.commit()
 
     def __validate_date_str(date_str: str, message: str):
         try:
-           datetime.strptime(date_str, "%Y-%m-%d")
+            datetime.strptime(date_str, "%Y-%m-%d")
         except ValueError:
-           raise APIException(message, 401)
-# Return fuction
+            raise APIException(message, 401)
